==> hpc/4_globus_download.ipynb <==
import timeit
# records the time at this instant
# of the program
runtime_start = timeit.default_timer()

# 4 Download raw output from HPC to CyberGIS-Jupyter for Water using Globus

Using this notebook, the user will be able to download the raw outputs on CyberGIS-Jupyter for Water from HPC. To enable this data trasnfer, CyberGIS-Compute integrates Globus service.

## 4_1 Prelimary step

### 4_1_1 Enter SUMMA job id

<div class="alert alert-block alert-danger">
<b> Enter SUMMA job id:  </b> Enter SUMMA job id created by notebook 2. You can find it in the cell 36 in the section 2_7_3. </div>

# the summa job to download
SUMMA_JOB_ID = '1651834763ORCy'

### 4_1_2 Import libraries

import os
import tempfile
import time
import json
# import CyberGIS-Compute Service SDK
from job_supervisor_client import *

## 4_2 Set up paths to the download folder

workspace = "./workspace"
workspace = os.path.abspath(workspace)
model_folder = "globus_" + str(int(time.time()))
model_folder = os.path.join(workspace, model_folder)
! mkdir -p {model_folder}
print(model_folder)

Note, this notebook currently can only work with the 'expanse' hpc (and not 'keeling').

job_folder_name="SUMMA_{}".format(SUMMA_JOB_ID)
jupyter_output_path=os.path.join(model_folder, job_folder_name)
!mkdir -p {jupyter_output_path}
jupyter_output_rel_path = jupyter_output_path.replace("/home/jovyan/work/", "")
data = dict(jupyter_user=os.environ["JUPYTERHUB_USER"], 
            hpc="expanse", 
            job_folder_name=job_folder_name, 
            jupyter_output_path=jupyter_output_rel_path)
data_path = os.path.join(model_folder, "globus.json")
with open(data_path, 'w') as outfile:
    json.dump(data, outfile)
    

! cat {data_path}

## 4_3 Communicate with CyberGIS-Compute Service SDK

### 4_3_1 Create session and upload model folder

#session = Session("globus", url="127.0.0.1", port=3000, isJupyter=True, protocol='HTTP')
session = Session("globus", isJupyter=True)

# upload model folder containing in.txt
job = session.job()
job.upload(model_folder)

### 4_3_2 Submit and monitor the job

job.submit()

# monitor job execution
job.events(liveOutput=True)

### 4_3_3 Locate the downloaded raw outputs on CyberGIS-Jupyter for Water

In this step, the users can find the location of the raw outputs on CyberGIS-Jupyter for Water that were downloaded from HPC using Globus. Then, they can navigate to this folder and do furthur calcualtions on the raw outputs as they wish or donwload necessary files to their local machines. 

!echo {jupyter_output_path}

Show the size and dicrectoy of downloaded raw outputs on CyberGIS-Jupyter for Water 

!du {jupyter_output_path} -h

# Print the time spent running the entire notebook.
# records the time at this instant 
# of the program
runtime_end = timeit.default_timer()

# printing the execution time by subtracting
# the time before the function from
# the time after the function
print(int(runtime_end-runtime_start), ' seconds')
